requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic[email]
pandas
numpy
scikit-learn
joblib
lightgbm
python-multipart
openpyxl
celery
redis
uuid

# Authentication and security
fastapi-users[sqlalchemy]>=12.0.0
fastapi-users[oauth]>=12.0.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
python-jose[cryptography]

# Web and templates
python-multipart

# Additional utilities
aiofiles
httpx
